Aigerim_Tolegenkyzy_task_11/requirements.txt
fastapi
uvicorn
celery
redis
python-jose[cryptography]
passlib[bcrypt]
python-multipart
